services/verification/requirements.txt
c2pa-python==0.4.0
cryptography==41.0.7
PyO3==0.20.0
pydantic>=2.0.0
pytest>=7.0.0
